[WIP] TF Refactor Mark II (#33)
See various README.md files for high level explanations of the moving parts.

As of right now infra/bootstrap and infra/singletons are correct. That is, there is a live TF state file in the s3 bucket which is up to date and tf plan does not want to change anything. There's a little bit left to do with app/stage and app/prod to bring them up to this same level, and nothing yet to do for infra/prod and infra/stage since we haven't created those things yet.

Once the state for each separate TF composition is up to date, we can merge this and remove the old original TF state file from the S3 bucket (which will then be superseded by the new split up ones)