chore: start decision log (#56)
QwerTech got me thinking about the decisions we're making and how we're not saving very much context on why we're making the decisions we've made. I thought a simple doc like this might be helpful way to save this context. Writing them out might also helps to question decisions that were right at the time and aren't anymore.