Add module for BigQuery data transfer configs
Code is WIP. Module is based on code I manually tested, but I have not tested it as part of this entire Terraform condig. Also, there are a few spots marked with TODOs. One is an opportunity to clean the code. The other is where we'll need to finish using the module (in both `stage` and `prod`) by providing the input parameters by integrating with the secrets store we're using.