feat: add BI dashboards project resources (#93)
Adds a Terraform module for the BI dashboards project and uses it in staging. We use BigQuery in GCP to store a copy of the data. We have one transfer job per batch of 5 tables. See comments in code for why. The code could be made more succinct/elegant by using more Terraform features to derive the batch info from a list of table names, but we're doing it this way so that even though it's more verbose, it should be easier for Terraform beginners to understand and edit in the future.

---------

Co-authored-by: rob salmond <rob@salmond.ca>