Add data monitoring alert
Also adds log based metric that powers the alert and the email address the notification channel of the alert uses when it fires. Also adds a variable to our BI dashboards module to specify email so the module can be used for prod too (we'll use a different email for prod alerts). Bit of a refactor for the `provider` block in the module since I'm using a new .tf file instead of adding it to bigquery.tf.